v2.5.2

+++ app/src/main/java/com/yoshione/fingen/utils/ScreenUtils.java
@@ -11,7 +11,7 @@ import android.util.DisplayMetrics;
 public class ScreenUtils {
 
 
-    public static int dpToPx(int dp, Context context) {
+    public static int dpToPx(float dp, Context context) {
         DisplayMetrics displayMetrics = context.getResources().getDisplayMetrics();
         return Math.round(dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT));
     }
